Update awake.lua (#19)
The version of awake distributed with the official SD card at the moment of writing is marked 2.4.0 but is significantly different from the one that is live one on GitHub from @tehn that is about 60 lines shorter, e.g. it lacks completely the management of MIDI start and stop messages. By marking the live version as a 2.4.1 revision we can avoid the ambiguity for other users like me.